Moving MaxResultSizeExceeded to funcx.utils.errors
