+++ funcx_sdk/funcx/tests/test_result_size.py
@@ -1,7 +1,6 @@
 import pytest
 
-from funcx.utils.errors import TaskPending
-from funcx_endpoint.executors.high_throughput.funcx_worker import MaxResultSizeExceeded
+from funcx.utils.errors import MaxResultSizeExceeded, TaskPending
 
 
 def large_result_producer(size) -> str:
