+++ smoke_tests/tests/test_s3_indirect.py
@@ -1,6 +1,6 @@
 import pytest
 
-from funcx_endpoint.executors.high_throughput.funcx_worker import MaxResultSizeExceeded
+from funcx.utils.errors import MaxResultSizeExceeded
 
 
 def large_result_producer(size: int) -> str:
